Add support for deleting resources
The --delete option can be used to delete the objects inside the
manifests. After each and every deletion the system will wait for the
deletion to actually happen. The idea behind is that this allows for
removing CRs and their CRDs in one pass, while waiting for the
finalizers to happen.